services/auth_service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Authentication service"
requires-python = ">=3.13"

dependencies = [
    "fastapi>=0.125.0",
    "uvicorn[standard]>=0.38.0",
    "sqlalchemy>=2.0.45",
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[argon2,bcrypt]>=1.7.4",
    "argon2-cffi>=25.1.0",
    "python-multipart>=0.0.21",
    "python-dotenv>=1.2.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "psycopg[binary]>=3.3.2",
]
